+++ packages/web/hooks/src/use-global-config/index.ts
