+++ packages/web/hooks/src/use-namespace/index.ts
