+++ packages/web/hooks/src/use-prevent-global/index.ts
