+++ packages/web/hooks/src/use-same-target/index.ts
