+++ packages/web/hooks/tests/use-deprecated.test.ts
@@ -2,7 +2,7 @@ import { computed, defineComponent, nextTick } from 'vue'
 import { mount } from '@vue/test-utils'
 import { afterEach, describe, expect, it, vi } from 'vitest'
 import { debugWarn } from '@element-plus/utils'
-import { useDeprecated } from '../use-deprecated'
+import { useDeprecated } from '../src'
 
 const AXIOM = 'Rem is the best girl'
