+++ packages/web/hooks/tests/use-locale.test.tsx
@@ -3,8 +3,7 @@ import { mount } from '@vue/test-utils'
 import { afterEach, beforeEach, describe, expect, it } from 'vitest'
 import Chinese from '@element-plus/locale/src/lang/zh-cn'
 import English from '@element-plus/locale/src/lang/en'
-import { buildTranslator, useLocale } from '../use-locale'
-import { provideGlobalConfig } from '..'
+import { buildTranslator, provideGlobalConfig, useLocale } from '../src'
 import type { Language } from '@element-plus/locale'
 import type { ComponentPublicInstance, PropType } from 'vue'
 import type { VueWrapper } from '@vue/test-utils'
