+++ packages/web/hooks/tests/use-model-toggle.test.tsx
@@ -2,7 +2,7 @@ import { defineComponent, nextTick, reactive, ref } from 'vue'
 import { mount } from '@vue/test-utils'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { useModelToggle, useModelToggleProps } from '../use-model-toggle'
+import { useModelToggle, useModelToggleProps } from '../src'
 import type { VueWrapper } from '@vue/test-utils'
 
 const AXIOM = 'Rem is the best girl'
