+++ packages/web/hooks/tests/use-timeout.test.ts
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-import { useTimeout } from '../use-timeout'
+import { useTimeout } from '../src'
 
 const _mount = (cb: () => void) => {
   return mount({
